ZF-10006: Zend_Captcha_Image::_gc() removes only captcha files identified by their suffix
git-svn-id: http://framework.zend.com/svn/framework/standard/trunk@22589 44c647ce-9c0f-0410-b52a-842ac1e357ba

+++ library/Zend/Captcha/Image.php
@@ -576,10 +576,14 @@ class Image extends Word
             // safety guard
             return;
         }
+        $suffixLength = strlen($this->_suffix);
         foreach (new \DirectoryIterator($imgdir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
                 if ($file->getMTime() < $expire) {
-                    unlink($file->getPathname());
+                    // only deletes files ending with $this->_suffix
+                    if (substr($file->getFilename(), -($suffixLength)) == $this->_suffix) {
+                        unlink($file->getPathname());
+                    }
                 }
             }
         }
